requirements.txt
flask==2.0.1
pdf2docx==0.5.6
python-docx==0.8.11
Werkzeug==2.0.1
tqdm==4.62.0
PyPDF2==2.10.5 
pikepdf==8.8.0
ghostscript==0.7
Pillow==10.2.0
pdf2image==1.16.3
img2pdf==0.5.1
reportlab==4.1.0
pdfrw==0.4.0
pdfplumber==0.10.3
tabula-py==2.9.0
pytesseract==0.3.10
pytest==7.4.0 